Normalise type-parsing routines and make arglist matching work
We now correctly match types in parsed arglists even in the case of
slot accesses.  Function calls and array accesses are still not
handled (we just pass UNION-TYPE through).  But, we do get some more
specialisation.  For example with some type definitions:

 type foo
   type(bar) :: slot
 end type foo

 type bar
   real, dimension(:), pointer :: n=>null()
 end type bar

and some code:

  type(foo) :: foo
  integer, dimension(10) :: x

  call some_function(x%slot, x%slot%n, x(1))

we derive the types of the arguments as:

 type(bar); real, dimension(:), pointer; UNION-TYPE

which is significantly better than the previous:

 UNION-TYPE; UNION-TYPE; UNION-TYPE