Speed up f90-clean-comments
Rather than moving through the buffer one character at a time, search
forward for a possible comment start (a ! character) and then use
parse-partial-sexp to decide if we're actually in a comment or not.

It turns out most of the time in parsing was actually cleaning
comments, so this makes the code run significantly faster overall.