Speed up f90-split-arglist
Previously we inserted the arglist into a temporary buffer and split
it there.  Since we're just using that to loop over the characters in
the string one at a time, do it in place by looping across the
characters of the array.  This latter method takes around 70% less
time than the former.  Since we split arglists thousands of times when
parsing a source tree, this is a worthwhile optimisation.